Adds "Insert a new record" feature
This commit adds the feature to insert new records. The new screen re-uses the use case where the latest record is shown, re-using the `HomeScreenInteractor`. Besides, when the new record is added, the metrics shown in the home screen are updated.

Still uses a mock repository and the values are stored in memory.