Adds calibration info when inserting new records
Sometimes the scales are not accurate, and the values expected and seen don’t match. In order to get a more accurate reading, a reference weight (an object which weight is known) can be used to see what’s the error seen in the scale. Those values can be specified as part as the `NewRecord` screen.