Moves the calibration logic outside the ViewController
Following the principle that the ViewController should be as dumb as possible, the business logic where the calibration is applied has been moved to the Presenter. Being that said, I’m not entirely sure that’s the best place. The other place it seems to fit is in the Interactor, since it is close to the business logic.