Adds water tracking to the list of the metrics
Water is an interesting metric that might provide a more accurate picture along with _muscle_ and _fat_. In order to add water tracking, the core data model needs to be updated (add a new field) and the old model migrated to the new one. Luckily, this can be done with the lightweight migration mechanism since Core Data can infer the migration model.