UIGraphView component addition
UIGraphView is a view component created to display values in a two dimensional graphs. At the moment, the only graphs supported are those where the x axis is equally distributed and the y axis is a proportional value. It displays a graph of points connected with a line. It comes with an Example project as well.

In the future I expect this UIGraphView to evolve and support different type of two-dimensional graphs (dispersion graphs, for instance).