Fixes an issue when displaying the current week on the first day of the week
The bug was caused when searching all the records within an interval. The time of the supplied hours was not reset to zero and, hence, the records of the same day but with a time previous of the today weren't shown. It has been fixed by addding to the presenter the logic to set the `Date` time to the beginning of the day.