Saves the record information in HealthKit
This commit adds HealthKit integration. When a new record is stored it'll be stored in HealthKit as well (if HealthKit is available in the device). The information stored is Weight, Height, Lean Body Mass, Body Fat Percentage and Body Mass Index.

The logic that stores the information in HealthKit can be found in `CreateNewRecord`.